Fix to compile on Linux

+++ Sources/Client/FTFont.cpp
@@ -18,7 +18,9 @@
 
  */
 
+#include <cmath>
 #include <ft2build.h>
+
 #include FT_FREETYPE_H
 #include FT_BITMAP_H
 
@@ -206,7 +208,7 @@ namespace spades {
 				binSize <<= 1;
 			}
 
-			baselineY = std::floorf(height * 1.f);
+			baselineY = std::floor(height * 1.f);
 
 			rendererIsLowQuality = dynamic_cast<draw::SWRenderer *>(renderer);
